Add SSD1306 constructor with implied address
Except for non-standard devices, the address value and the display size
change in unisom.  There are exactly two pairs of these values and
therefore requiring the caller to specify both the address and size
in the construtor means more work, more possible errors, and knowledge
on the side of the user of the library which can be avoided to be
hardcoded.

This patch introduces a second constructor which takes only the size
parameter and derives the address parameter.